refactor(checker): split constructor checks out of ambient_signature_checks
The TS2366 strictNullChecks gate added lines to ambient_signature_checks.rs,
tipping it from 1999 to 2008 lines — over the 2000-line architecture limit
(arch-size runs only in merge_group, so PR-head CI didn't catch it). Move the
constructor-declaration checks (check_constructor_declaration{,_with_request},
check_jsdoc_constructor_tags) into a sibling module ambient_constructor_checks.rs.
Pure refactor: same impl, same behavior (verified TS2366 flips + constructor
diagnostics unchanged). Main file now 1601 lines.
